Parcial2.ReadLabyrinth

Description

Asume que el laberinto será un archivo de texto, en donde se definirán una serie de nodos y una matriz de adyacencia. Para ello, se utilizará el siguiente formato:

a. Línea 1: un número N que indica cuantos nodos habrá en el laberinto. b. Siguientes N líneas: vi1, vi2, ….,viN (cada vij es un 0 o 1, donde 0 representa que no existe adyacencia entre el nodo de la i-ésima línea y en nodo de la j-ésima columna). c. Siguiente línea: vI, vF (dos números, donde vI representa el nodo de partida del agente robótico y vF representa el nodo al cual tiene que llegar). d. Siguientes N líneas: x,y (donde xi,yi representan las coordenadas donde se encuentra en nodo i-ésimo).

Documentation

data LabyrinthDescription

Constructors

LabyrinthDescription Int [(Int, Int)] (Int, Int) [(Int, Int)]

buildConnections :: [[Bool]] -> [(Int, Int)]

parseLabyrinth :: [Char] -> Either [String] LabyrinthDescription

labyrinthFile :: GenParser Char st LabyrinthDescription

readComma :: ParsecT [Char] u Identity ()

readPair :: Read a => GenParser Char st (a, a)